Support MTGJSON UUID Cache - Prevents UUIDs of cards from changing once they have been assigned - Improves trust in our UUID system
