+++ mtgjson5/providers/__init__.py
@@ -12,7 +12,6 @@ from .github.github_card_sealed_products import GitHubCardSealedProductsProvider
 from .github.github_decks import GitHubDecksProvider
 from .github.github_mtgsqlite import GitHubMTGSqliteProvider
 from .github.github_sealed import GitHubSealedProvider
-from .github.github_uuid_cache import GitHubUuidCacheProvider
 from .manapool.manapool_prices import ManapoolPricesProvider
 from .mtgwiki.secret_lair import MtgWikiProviderSecretLair
 from .multiversebridge import MultiverseBridgeProvider
@@ -20,5 +19,6 @@ from .scryfall.monolith import ScryfallProvider
 from .scryfall.orientation_detector import ScryfallProviderOrientationDetector
 from .scryfall.set_language_detector import ScryfallProviderSetLanguageDetector
 from .tcgplayer import TCGPlayerProvider
+from .uuid_cache import UuidCacheProvider
 from .whats_in_standard import WhatsInStandardProvider
 from .wizards import WizardsProvider
